requirements.txt
Flask==2.1.1            # Flask for the web framework
pandas==1.5.2           # Pandas for data handling
numpy==1.23.2           # NumPy for numerical operations
statsmodels==0.13.2     # Statsmodels for ARIMA modeling
scikit-learn==1.0.2     # Scikit-learn (if you're using it for preprocessing or model evaluation)
joblib==1.1.0           # Joblib for saving/loading the ARIMA model
matplotlib==3.5.2       # Matplotlib for plotting (if you're visualizing data or results)
flask-cors==3.1.1       # Flask CORS for handling cross-origin requests (if needed)
python-dotenv==1.21.0
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==1.0.0
Flask-Login==0.6.2
psycopg2==2.9.3
Flask-Migrate==3.1.0